refactor(adr-061): physically migrate severity/exit_decision/contract_coverage_exit to abicheck/policy/
Moves three of policy's four `legacy_paths` entries from flat
abicheck/<name>.py to the real abicheck/policy/ subpackage (mirroring the
existing model/compare/storage precedent), leaving thin static back-compat
shims at the old flat paths so every existing import path keeps resolving
to the identical object.

analysis_assurance.py (the fourth legacy_paths entry) is deliberately left
at its flat path: its own architecture/debt.yaml no_growth entry already
states it "cannot move safely without a behavior-preserving vertical
slice," and check_architecture.py's debt-exemption gate mechanically
enforces exactly that for any PR that would relocate a debt-tracked file's
path without a matching base-revision entry.

checker_policy.py/contract_gating.py/reclassify.py/contract_coverage_ledger.py
stay unclassified in architecture/modules.yaml, added to
public_root_surfaces instead of a legacy_paths layer -- each is imported
too widely across every responsibility layer (including frontends, which
may not import compare/policy) to safely receive a layer classification;
two of them already document themselves as deliberately-unclassified
cross-layer leaves.

See the changelog fragment for the full account, including two mypy
attr-defined re-export fixes and two test files updated to reference the
real module for a private helper the shim's __all__ deliberately excludes.
